refactored CMake and file structure for play mp4 from beginning sample
