+++ samples/play_mp4_from_beginning/PlayMp4VideoFromBeginning.h
@@ -1,24 +1,24 @@
-#include <PipeLine.h>
-#include "Mp4ReaderSource.h"
-#include "ImageViewerModule.h"
 #include "ColorConversionXForm.h"
 #include "H264Decoder.h"
+#include "ImageViewerModule.h"
+#include "Mp4ReaderSource.h"
 
+#include <PipeLine.h>
 
 class PlayMp4VideoFromBeginning {
 public:
-    PlayMp4VideoFromBeginning(); 
-    bool setUpPipeLine();
-    bool startPipeLine();
-    bool stopPipeLine();
-    bool flushQueuesAndSeek();
-    bool testPipeLineForFlushQue();
-    bool testPipeLineForSeek();
+  PlayMp4VideoFromBeginning();
+  bool setUpPipeLine(const std::string &videoPath);
+  bool startPipeLine();
+  bool stopPipeLine();
+  bool flushQueuesAndSeek();
+
+
+  boost::shared_ptr<Mp4ReaderSource> mp4Reader;
+  boost::shared_ptr<ImageViewerModule> imageViewerSink;
+  boost::shared_ptr<H264Decoder> decoder;
+  boost::shared_ptr<ColorConversion> colorchange;
 
 private:
-    PipeLine pipeLine;
-    boost::shared_ptr<Mp4ReaderSource> mp4Reader;
-    boost::shared_ptr<ImageViewerModule> imageViewerSink;
-    boost::shared_ptr<H264Decoder> decoder;
-    boost::shared_ptr<ColorConversion> colorchange;
+  PipeLine pipeLine;
 };
